Make a SummaryView and ReviewView inheritance to BaseView
